requirements.txt
PyYAML==5.4.1

numpy
scipy
joblib
pandas
cachetools
aiohttp
